pyproject.toml
[project]
name = "agent-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.74.0",
    "python-dotenv>=1.2.2",
]

[dependency-groups]
dev = ["ruff>=0.6.0"]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["B008", "E501"]
